codejangler/pyproject.toml
[project]
name = "codejangler"
version = "0.1.0"
description = "CLI tool for optimizing GitHub PRs using CrewAI"
authors = [
    {name = "Jason DeLorenzo", email = "delorenj@gmail.com"}
]
dependencies = [
    "click",
    "crewai",
    "gitpython",
    "PyGithub"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "pre-commit"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
